fix(game): await getRandomWordPair to ensure proper word assignment in room creation
Previously, getRandomWordPair() was being called without await, causing a Promise object to be stored instead of a resolved word pair. This led to blank word values on the frontend. The route is now async and correctly awaits the result.
