+++ src/public/create-game.html
@@ -19,6 +19,12 @@
         <p class="text-muted">Share this code with friends to join.</p>
       </div>
 
+    <div id="word-pair-container" class="mt-4 d-none">
+        <p class="fw-bold text-primary">Assigned Word Pair:</p>
+        <p><strong>Civilian:</strong> <span id="civilian-word"></span></p>
+        <p><strong>Undercover:</strong> <span id="undercover-word"></span></p>
+      </div>
+
     <button id="join-game-button" class="btn btn-secondary mt-4">Join Game</button>
 
     <div id="join-game-container" class="mt-3 d-none">
